backend/requirements.txt
djangorestframework==3.12.2
django-environ==0.4.5
boto3==1.9.252
botocore==1.12.252
Django==3.1.6
django-storages==1.7.2
docutils==0.15.2
gunicorn==20.0.4
jmespath==0.9.4
psycopg2==2.8.6
python-dateutil==2.8.0
pytz==2019.3
s3transfer==0.2.1
six==1.12.0
sqlparse==0.4.1
urllib3==1.25.6
djangorestframework_simplejwt==4.6.0
django-filter==2.4.0
Pillow==8.2.0
django-model-controller==0.4.1
drf-writable-nested==0.6.2
django-extensions==3.1.3
ipdb==0.13.2

